app/requirements.txt
Django==4.1
gunicorn==20.1.0
psycopg2-binary==2.9.1
djangorestframework==3.13.1
django-extensions==3.1.3
requests==2.28.1
django-cors-headers==3.13.0